+++ packages/stable/src/index.ts
@@ -49,7 +49,6 @@ export { RawAPI } from './api/raw/rawApi';
 export { RawRowsAPI } from './api/raw/rawRowsApi';
 export { RawTablesAPI } from './api/raw/rawTablesApi';
 export { EntityMatchingApi } from './api/entityMatching/entityMatchingApi';
-export { AnnotationsAPI } from './api/annotations/annotationsApi';
 export { RelationshipsApi } from './api/relationships/relationshipsApi';
 export { SecurityCategoriesAPI } from './api/securityCategories/securityCategoriesApi';
 export { SequenceRowsAPI } from './api/sequences/sequenceRowsApi';
